Added accountInfo template, moved controller logic and added new method, modified header
