+++ src/main/resources/templates/accountInfo.html
@@ -1,0 +1,12 @@
+<!DOCTYPE html>
+<html xmlns:th="http://www.thymeleaf.org" lang="en">
+<head>
+    <meta charset="UTF-8">
+    <title>Account Information</title>
+</head>
+<body>
+<div th:replace="~{fragments/header :: header}">Header </div>
+
+
+</body>
+</html>
