+++ src/main/resources/templates/fragments/header.html
@@ -11,7 +11,7 @@
       <a class="navbar-brand" th:href="@{/home}">Home</a>
     </div>
     <ul class="nav navbar-nav">
-      <li><a class="nav-link" th:href="@{/customer}">Your Account</a></li>
+      <li><a class="nav-link" th:href="@{/accountInfo}">Your Account</a></li>
       <li><a class="nav-link" th:href="@{/reviews}">Your Reviews</a></li>
     </ul>
   </div>
