+++ src/main/resources/templates/product.html
@@ -5,8 +5,8 @@
   <title>Product Details</title>
 </head>
 <body>
+<div th:replace="~{fragments/header :: header}">Header </div>
 
-  <a th:href="@{/home}">Back to Home</a>
   <h1>Product Details</h1>
 
   <ul>
